Ensure multipart resources always get unique filenames in the bundle
Resource IDs must be unique per “level” — ie. within the “top” level, then within the resources of each multipart resource, and so on.

Previously, adding the same multipart resource more than once would result in the resources of each multipart resource to have the same filename into the bundle.

This wasn’t a problem as all resources were still distinct, but the fact this worked reliably was predicated on implementation details—specifically, the fact that we’d build a flat list of resources-to-read and sequentially matching them to returned tar entries.

This commit assigns distinct filenames to all resources to make things more reliable.

Change-type: minor